Added files for account and data DAO
